# doc/technical/openapi.yaml
openapi: 3.1.0
info:
  title: 'SoManAgent API'
  version: 1.0.0
  description: 'Versioned OpenAPI contract for the SoManAgent backend. This specification is spec-first: operations may be documented ahead of implementation when they explicitly set `x-somanagent-implemented` to `false`.'
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /api/logs/events:
    get:
      tags:
        - Logs
      summary: 'List log events'
      operationId: listLogEvents
      responses:
        200:
          $ref: '#/components/responses/Ok'
    post:
      tags:
        - Logs
      summary: 'Ingest a log event'
      operationId: ingestLogEvent
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        202:
          $ref: '#/components/responses/Accepted'
  '/api/logs/occurrences/{id}':
    get:
      tags:
        - Logs
      summary: 'Get a log occurrence'
      operationId: getLogOccurrence
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/logs/occurrences/{id}/status':
    patch:
      tags:
        - Logs
      summary: 'Update log occurrence status'
      operationId: updateLogOccurrenceStatus
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
    ArbitraryJsonBody:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArbitraryJsonObject'
...
  responses:
    Ok:
      description: 'Successful response'
...
    Accepted:
      description: 'Request accepted for asynchronous processing'
...
  schemas:
...
    ArbitraryJsonObject:
      type: object
      additionalProperties: true
...